Move functions around in stats package
